We are getting it boisssss

+++ src/main/java/de/szut/lf8_project/lf8/project/ProjectController.java
@@ -2,7 +2,7 @@ package de.szut.lf8_project.lf8.project;
 
 import de.szut.employees.EmployeeAPI;
 import de.szut.employees.dto.EmployeeResponseDTO;
-import de.szut.lf8_project.exceptionHandling.InvalidDataException;
+import de.szut.lf8_project.exceptionHandling.ResourceNotFoundException;
 import de.szut.lf8_project.lf8.project.dto.CreateProjectDto;
 import de.szut.lf8_project.lf8.project.dto.GetProjectDto;
 import de.szut.lf8_project.lf8.timemanagement.TimeManagementEntity;
@@ -62,10 +62,10 @@ public class ProjectController {
 
                 timeManagementService.create(timeManagementEntity);
 
-                entity.getEmployees().add(timeManagementEntity);
+                entity.getEmployees().add(timeManagementEntity.getId());
             }
 
-            service.create(entity);
+            entity = service.create(entity);
         }
 
         return new ResponseEntity<>(mapper.mapToGetProjectDto(entity), HttpStatus.CREATED);
@@ -81,12 +81,18 @@ public class ProjectController {
                     content = @Content),
             @ApiResponse(responseCode = HTTPCodes.NOT_AUTHORIZED, description = "not authorized",
                     content = @Content),
+            @ApiResponse(responseCode = HTTPCodes.NOT_FOUND, description = "the project was not found",
+                    content = @Content),
             @ApiResponse(responseCode = HTTPCodes.INTERNAL_SERVER_ERROR, description = "internal server error",
                     content = @Content)
     })
     @GetMapping (path = "{projectId}")
     public ResponseEntity<GetProjectDto> getById(@PathVariable final Long projectId) {
         final var entity = this.service.readById(projectId);
+
+        if(entity == null)
+            throw new ResourceNotFoundException("The project was not found");
+
         final GetProjectDto dto = this.mapper.mapToGetProjectDto(entity);
         return new ResponseEntity<>(dto, HttpStatus.OK);
     }
@@ -121,6 +127,9 @@ public class ProjectController {
     })
     @DeleteMapping("{projectId}")
     public void DeleteProject(@PathVariable Long projectId) {
+        if(service.readById(projectId) == null)
+            throw new ResourceNotFoundException("The project was not found");
+
         service.deleteProjectById(projectId);
     }
 
@@ -133,7 +142,8 @@ public class ProjectController {
             @ApiResponse(responseCode = HTTPCodes.BAD_REQUEST, description = "invalid JSON posted",
                     content = @Content),
             @ApiResponse(responseCode = HTTPCodes.NOT_AUTHORIZED, description = "not authorized",
-
+                    content = @Content),
+            @ApiResponse(responseCode = HTTPCodes.NOT_FOUND, description = "the project was not found",
                     content = @Content),
             @ApiResponse(responseCode = HTTPCodes.INTERNAL_SERVER_ERROR, description = "internal server error",
                     content = @Content)
@@ -142,8 +152,12 @@ public class ProjectController {
     public ResponseEntity<HashMap<Long, String>> getAllEmployeesByProjectId(@RequestHeader("Authorization") String authToken, @PathVariable final Long projectId) throws IOException {
         HashMap<Long, String> employeeResponseDTOList = new HashMap<>();
         final ProjectEntity entity = this.service.readById(projectId);
-        for (TimeManagementEntity employeeID : entity.getEmployees()) {
-            EmployeeResponseDTO foundEmployee = EmployeeAPI.getInstance().findEmployeeById(employeeID.getEmployeeId(), authToken);
+
+        if(entity == null)
+            throw new ResourceNotFoundException("The project was not found");
+
+        for (TimeManagementEntity employee : timeManagementService.readAllById(entity.getEmployees())) {
+            EmployeeResponseDTO foundEmployee = EmployeeAPI.getInstance().findEmployeeById(employee.getEmployeeId(), authToken);
             employeeResponseDTOList.put(foundEmployee.getId(), String.format("%s %s", foundEmployee.getFirstName(), foundEmployee.getLastName()));
         }
 
@@ -160,6 +174,8 @@ public class ProjectController {
                     content = @Content),
             @ApiResponse(responseCode = HTTPCodes.NOT_AUTHORIZED, description = "not authorized",
                     content = @Content),
+            @ApiResponse(responseCode = HTTPCodes.NOT_FOUND, description = "The project was not found",
+                    content = @Content),
             @ApiResponse(responseCode = HTTPCodes.INTERNAL_SERVER_ERROR, description = "internal server error",
                     content = @Content)
     })
@@ -168,7 +184,7 @@ public class ProjectController {
     public ResponseEntity<GetProjectDto> addEmployeeToProject(@RequestHeader("Authorization") String authToken, @PathVariable Long projectId, @RequestBody @Valid AddEmployeeToProjectDto dto) throws IOException {
         ProjectEntity entity = service.readById(projectId);
         if(entity == null)
-            throw new InvalidDataException("The project was not found!");
+            throw new ResourceNotFoundException("The project was not found!");
 
         HelperFunctions.checkEmployee(dto.getEmployeeId(), entity, dto.getStartDate(), dto.getEndDate(), authToken, timeManagementService);
 
@@ -180,7 +196,7 @@ public class ProjectController {
 
         timeManagementService.create(tmEntity);
 
-        entity.getEmployees().add(tmEntity);
+        entity.getEmployees().add(tmEntity.getId());
 
         return new ResponseEntity<>(mapper.mapToGetProjectDto(service.create(entity)), HttpStatus.CREATED);
     }
@@ -188,31 +204,31 @@ public class ProjectController {
     @Operation(summary = "Updates the Project")
     @ApiResponses(value = {
             @ApiResponse(responseCode = HTTPCodes.SUCCESSFUL, description = "update successful", content = @Content),
-            @ApiResponse(responseCode = HTTPCodes.NOT_FOUND, description = "not found"),
+            @ApiResponse(responseCode = HTTPCodes.NOT_FOUND, description = "the project was not found"),
             @ApiResponse(responseCode = HTTPCodes.INTERNAL_SERVER_ERROR, description = "internal server error",
                     content = @Content)
     })
     @PutMapping(path = "{id}")
-    public ResponseEntity <GetProjectDto> update(@PathVariable final Long id, @RequestBody @Valid CreateProjectDto dto) {
+    public ResponseEntity <GetProjectDto> update(@RequestHeader("Authorization") String authToken, @PathVariable final Long id, @RequestBody @Valid CreateProjectDto dto) throws IOException {
         ProjectEntity projectEntity = service.readById(id);
+        mapper.mapCreateDtoToEntity(dto, authToken);
 
-        if (projectEntity == null) {
-            throw new InvalidDataException("Project not found");
-        }
+        if (projectEntity == null)
+            throw new ResourceNotFoundException("Project not found");
 
         if(dto.getEmployees() != null) {
+            Set<Long> processedEmployees = new HashSet<>();
             timeManagementService.deleteAllByProjectId(id);
 
             if(!dto.getEmployees().isEmpty()) {
-                Set<TimeManagementEntity> processedEmployees = new HashSet<>();
                 dto.getEmployees().forEach(i -> processedEmployees.add(
                         timeManagementService.create(
                                 new TimeManagementEntity(0L, projectEntity.getId(), i.getEmployeeId(), i.getStartDate(), i.getEndDate())
-                        )
+                        ).getId()
                 ));
-
-                projectEntity.setEmployees(processedEmployees);
             }
+
+            projectEntity.setEmployees(processedEmployees);
         }
 
         projectEntity.setEndDate(dto.getEndDate());
@@ -227,28 +243,31 @@ public class ProjectController {
 
     @Operation(summary = "Delete a projects employee")
     @ApiResponses(value = {
-            @ApiResponse(responseCode =  HTTPCodes.SUCCESSFUL, description = "employee was deleted", content = {
+            @ApiResponse(responseCode =  HTTPCodes.SUCCESSFUL, description = "Employee was deleted", content = {
                     @Content(mediaType = MediaTypes.JSON,
                             schema = @Schema(implementation = GetProjectDto.class))
             }),
-            @ApiResponse(responseCode = HTTPCodes.NOT_FOUND, description = "not found",
+            @ApiResponse(responseCode = HTTPCodes.NOT_FOUND, description = "Project not found",
                     content = @Content),
             @ApiResponse(responseCode = HTTPCodes.INTERNAL_SERVER_ERROR, description = "internal server error",
                     content = @Content)
     })
     @DeleteMapping (path = "{projectId}/employees/{employeeId}")
-    public ResponseEntity<GetProjectDto> deleteEmployeeFromProject(@PathVariable Long employeeId, @PathVariable Long projectId) throws IOException {
+    public ResponseEntity<GetProjectDto> deleteEmployeeFromProject(@PathVariable Long employeeId, @PathVariable Long projectId) {
         ProjectEntity project = service.readById(projectId);
         if(project == null)
-            throw new InvalidDataException("Project not found");
+            throw new ResourceNotFoundException("Project not found");
 
         if (project.getEmployees() == null)
-            throw new InvalidDataException("No employee in project");
+            return new ResponseEntity<>(mapper.mapToGetProjectDto(project), HttpStatus.OK);
 
-        if(project.getEmployees().contains(employeeId))
-            project.getEmployees().remove(employeeId);
-        else
-            throw new InvalidDataException("Employee could not be deleted");
+        TimeManagementEntity tmEntity = timeManagementService.readByEmployeeIdAndProjectId(employeeId, projectId);
+
+        if(tmEntity != null)
+            project.getEmployees().remove(tmEntity.getId());
+
+        timeManagementService.deleteByEmployeeIdAndProjectId(employeeId, projectId);
+
         return new ResponseEntity<>(mapper.mapToGetProjectDto(project), HttpStatus.OK);
     }
